+++ src/scraper/spider.py
@@ -2,8 +2,7 @@ import os
 
 from scrapy.http.response.html import HtmlResponse
 from scrapy.exceptions import CloseSpider
-from scrapy.linkextractors import LinkExtractor
-from scrapy.spiders import CrawlSpider, Rule
+from scrapy.spiders import CrawlSpider
 
 CRAWL_LIMIT = os.getenv("CRAWL_LIMIT", 100)
 OVERRIDE_SETTINGS = {
@@ -15,23 +14,9 @@ OVERRIDE_SETTINGS = {
 }
 
 
-class GinaCodySpider(CrawlSpider):
-    name = "ginacody"
-    start_urls = ["https://www.concordia.ca/ginacody.html/"]
+class ModifiedSpider(CrawlSpider):
+    name = "general"
     visited_urls = set()
-    rules = [
-        Rule(
-            LinkExtractor(
-                allow="ginacody",
-                deny=(
-                    r"^(?!https://www.concordia.ca).+",
-                    r"(https://www.concordia.ca/fr).+",
-                ),
-            ),
-            callback="parse_item",
-            follow=True,
-        )
-    ]
     crawl_limit = int(CRAWL_LIMIT)
 
     def parse_item(self, response: HtmlResponse) -> None:
@@ -63,7 +48,8 @@ class GinaCodySpider(CrawlSpider):
             "//meta[@name='robots']/@content"
         ).extract_first()
         if meta_robots_tag_content and (
-            "noindex" in meta_robots_tag_content or "nofollow" in meta_robots_tag_content
+            "noindex" in meta_robots_tag_content
+            or "nofollow" in meta_robots_tag_content
         ):
             return False
         return True
